misc/HtmlTestSuite/results/test-1374.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><svg xmlns="http://www.w3.org/2000/svg">
  <text>
    <tspan id="container">
      <tspan id="element"/>
    </tspan>
  </text>
</svg>

<script>
var element = document.getElementById('element');
var container = document.getElementById('container');

test(function() {
    element.style.fontSize = '10px';
    var player = element.animate([{baselineShift: '3em'}, {baselineShift: '5em'}], 10);
    player.pause();
    player.currentTime = 5;
    element.style.fontSize = '20px';
    assert_equals(getComputedStyle(element).baselineShift, '80px');
}, 'baselineShift responsive to style changes');

test(function() {
    container.style.baselineShift = 'sub';
    var player = element.animate([{baselineShift: 'inherit'}, {baselineShift: '20px'}], 4000);
    player.pause();

    player.currentTime = 1000;
    assert_equals(getComputedStyle(element).baselineShift, 'sub');

    container.style.baselineShift = 'super';
    assert_equals(getComputedStyle(element).baselineShift, 'super');

    container.style.baselineShift = '100px';
    assert_equals(getComputedStyle(element).baselineShift, '80px');

    container.style.baselineShift = 'sub';
    assert_equals(getComputedStyle(element).baselineShift, 'sub');
}, 'baselineShift responsive to inherited changes');
</script>
</body></html>
